+++ ome_zarr_neuro/enums.py
@@ -1,0 +1,14 @@
+from enum import Enum, auto
+
+class TransformType(Enum):
+    AFFINE_RAS = auto()
+    DISPLACEMENT_RAS = auto()
+
+
+class ImageType(Enum):
+    OME_ZARR = auto()
+    ZARR = auto()
+    NIFTI = auto()
+    UNKNOWN = auto()
+
+
